Refactored admin page to server-side rendering
- Fetching fallen data directly from service instead of API.
- Rendering data using FallenTable component.

BUG: [Server] Error: Maximum call stack size exceeded
