--- app/admin/page.module.scss
@@ -1,214 +1,0 @@
-@use '@/styles/_vars.scss' as *;
-
-.wrapper {
-    display: flex;
-    gap: 10px;
-    align-items: center;
-    width: 90%;
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 15px;
-    border-radius: 10px;
-    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
-
-    .searchContainer {
-        position: relative;
-        display: flex;
-        align-items: center;
-        width: 100%;
-
-        .searchIcon {
-            position: absolute;
-            right: 15px;
-            top: 50%;
-            transform: translateY(-50%);
-            pointer-events: none;
-        }
-
-        .searchInput {
-            color: $textColor;
-            border: 0.5px solid $lightblueBorder;
-            border-radius: 10px;
-            padding: 10px;
-            padding-right: 45px;
-            width: 100%;
-            font-size: $text-sm;
-            font-family: $font-family;
-
-            &::placeholder {
-                color: $placeHolderColor;
-            }
-        }
-    }
-
-    .filterButton {
-        background-color: white;
-        width: auto;
-        padding: 8px 20px;
-        gap: 5px;
-        border-radius: 10px;
-        font-size: $text-md;
-        font-family: $font-family;
-        border: 0.5px solid $lightblueBorder;
-        cursor: pointer;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        width: auto;
-    }
-}
-
-.modalOverlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.5);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.modal {
-    background-color: white;
-    padding: 20px 30px;
-    border-radius: 15px;
-    position: relative;
-    width: auto;
-    text-align: center;
-    min-height: 150px;
-    min-width: 300px;
-
-    h2 {
-        color: $titleColor;
-        margin-bottom: 15px;
-        font-size: $text-lg;
-    }
-
-    label {
-        margin-top: 15px;
-    }
-
-    select {
-        border: 0.5px solid $lightblueBorder;
-        border-radius: 8px;
-        padding: 8px;
-        font-size: 14px;
-        margin-right: 5px;
-    }
-
-    .applyButton {
-        color: white;
-        background-color: $buttonColor;
-        width: auto;
-        padding: 8px 15px;
-        border-radius: 10px;
-        font-size: $text-sm;
-        font-family: $font-family;
-        border: none;
-        cursor: pointer;
-        display: block;
-        margin-top: 15px;
-        margin-right: auto;
-    }
-
-    .closeButton {
-        position: absolute;
-        top: 12px;
-        right: 12px;
-        background-color: transparent;
-        border: none;
-        font-size: 18px;
-        cursor: pointer;
-        color: var(--lightGray);
-    }
-}
-
-.fallenTable {
-    width: 90%;
-    max-width: 1200px;
-    border-collapse: collapse;
-    margin: 0 auto;
-    margin-top: 20px;
-    border-radius: 10px;
-    overflow: hidden;
-    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
-
-    thead {
-        background-color: var(--lightGray);
-        color: $textColor;
-    }
-
-    tbody tr {
-        border-bottom: 1px solid var(--lightGray);
-    }
-
-    tbody tr:last-child {
-        border-bottom: none;
-    }
-
-    th,
-    td {
-        padding: 15px 15px;
-        text-align: center;
-        font-family: $font-family;
-        font-size: $text-sm;
-    }
-
-    .date {
-        color: $textColor;
-    }
-
-    .status {
-        border-radius: 16px;
-        padding: 2px 5px;
-        font-weight: 500;
-    }
-
-    .status.pending {
-        background-color: #FFF9C4;
-        color: #854D0E;
-    }
-
-    .status.approved {
-        background-color: #C8E6C9;
-        color: #2E7D32;
-    }
-
-    .status.rejected {
-        background-color: #FFCDD2;
-        color: #C62828;
-    }
-
-    .actions button {
-        background-color: transparent;
-        padding: 5px;
-        border: none;
-        cursor: pointer;
-    }
-}
-
-.modalEditFallen {
-    background-color: white;
-    padding: 10px;
-    border-radius: 15px;
-    position: relative;
-    text-align: center;
-    min-height: 150px;
-    min-width: 300px;
-    overflow: auto; 
-    max-height: 90vh; 
-
-
-    .closeButtonEditFallenModal {
-        position: absolute;
-        top: 12px;
-        right: 12px;
-        background-color: transparent;
-        border: none;
-        font-size: 18px;
-        cursor: pointer;
-        color: var(--lightGray);
-    }
-}
